+++ kart/completion.py
@@ -1,9 +1,10 @@
 import os
 import re
+import subprocess
 import sys
 from enum import Enum
 from pathlib import Path
-from typing import Optional, Tuple, Optional, Any,Dict, List
+from typing import Optional, Tuple, Optional, Any, Dict, List
 
 import click
 from click.shell_completion import _SOURCE_BASH, _SOURCE_ZSH, _SOURCE_FISH
@@ -18,12 +19,38 @@ class Shells(str, Enum):
     bash = "bash"
     zsh = "zsh"
     fish = "fish"
+    powershell = "powershell"
+    pwsh = "pwsh"
 
+
+_SOURCE_POWERSHELL = """
+Import-Module PSReadLine
+Set-PSReadLineKeyHandler -Chord Tab -Function MenuComplete
+$scriptblock = {
+    param($wordToComplete, $commandAst, $cursorPosition)
+    $Env:%(complete_var)s = "complete_powershell"
+    $Env:_KART_COMPLETE_ARGS = $commandAst.ToString()
+    $Env:_KART_COMPLETE_WORD_TO_COMPLETE = $wordToComplete
+    %(prog_name)s | ForEach-Object {
+        $commandArray = $_ -Split ":::"
+        $command = $commandArray[0]
+        $helpString = $commandArray[1]
+        [System.Management.Automation.CompletionResult]::new(
+            $command, $command, 'ParameterValue', $helpString)
+    }
+    $Env:%(complete_var)s = ""
+    $Env:_KART_COMPLETE_ARGS = ""
+    $Env:_KART_COMPLETE_WORD_TO_COMPLETE = ""
+}
+Register-ArgumentCompleter -Native -CommandName %(prog_name)s -ScriptBlock $scriptblock
+"""
 
 _completion_scripts = {
     "bash": _SOURCE_BASH,
     "zsh": _SOURCE_ZSH,
     "fish": _SOURCE_FISH,
+    "powershell": _SOURCE_POWERSHELL,
+    "pwsh": _SOURCE_POWERSHELL,
 }
 
 _invalid_ident_char_re = re.compile(r"[^a-zA-Z0-9_]")
@@ -109,6 +136,49 @@ def install_fish(*, prog_name: str, complete_var: str, shell: str) -> Path:
     return path_obj
 
 
+def install_powershell(*, prog_name: str, complete_var: str, shell: str) -> Path:
+    subprocess.run(
+        [
+            shell,
+            "-Command",
+            "Set-ExecutionPolicy",
+            "Unrestricted",
+            "-Scope",
+            "CurrentUser",
+        ]
+    )
+    result = subprocess.run(
+        [shell, "-NoProfile", "-Command", "echo", "$profile"],
+        check=True,
+        stdout=subprocess.PIPE,
+    )
+    if result.returncode != 0:
+        click.echo("Couldn't get PowerShell user profile", err=True)
+        raise click.exceptions.Exit(result.returncode)
+    path_str = ""
+    if isinstance(result.stdout, str):
+        path_str = result.stdout
+    if isinstance(result.stdout, bytes):
+        try:
+            # PowerShell would be predominant in Windows
+            path_str = result.stdout.decode("windows-1252")
+        except UnicodeDecodeError:
+            try:
+                path_str = result.stdout.decode("utf8")
+            except UnicodeDecodeError:
+                click.echo("Couldn't decode the path automatically", err=True)
+                raise click.exceptions.Exit(1)
+    path_obj = Path(path_str.strip())
+    parent_dir: Path = path_obj.parent
+    parent_dir.mkdir(parents=True, exist_ok=True)
+    script_content = get_completion_script(
+        prog_name=prog_name, complete_var=complete_var, shell=shell
+    )
+    with path_obj.open(mode="a") as f:
+        f.write(f"{script_content}\n")
+    return path_obj
+
+
 def install(
     shell: Optional[str] = None,
     prog_name: Optional[str] = None,
@@ -136,6 +206,11 @@ def install(
             prog_name=prog_name, complete_var=complete_var, shell=shell
         )
         return shell, installed_path
+    elif shell in {"powershell", "pwsh"}:
+        installed_path = install_powershell(
+            prog_name=prog_name, complete_var=complete_var, shell=shell
+        )
+        return shell, installed_path
     else:
         click.echo(f"Shell {shell} is not supported.")
         raise click.exceptions.Exit(1)
@@ -156,7 +231,7 @@ def install_callback(ctx: click.Context, param: click.Parameter, value: Any) -> 
 @click.shell_completion.add_completion_class
 class PowerShellComplete(click.shell_completion.ShellComplete):
     name = Shells.powershell.value
-    source_template = COMPLETION_SCRIPT_POWER_SHELL
+    source_template =_SOURCE_POWERSHELL 
 
     def source_vars(self) -> Dict[str, Any]:
         return {